refactor(dashboard): extract name generator to dedicated utility module
Move adjectives/nouns arrays, generateName(), and toSlug() from
page.tsx into lib/name-generator.ts. Reduces page component from
~580 lines to ~290 lines. Arrays are now module-level constants
(allocated once instead of per-call).
