+++ packages/dashboard/src/lib/name-generator.ts
@@ -1,0 +1,298 @@
+const adjectives = [
+  "Fast",
+  "Smart",
+  "Bright",
+  "Bold",
+  "Swift",
+  "Keen",
+  "Calm",
+  "Cool",
+  "Sharp",
+  "Prime",
+  "Brave",
+  "Clear",
+  "Deep",
+  "Fair",
+  "Grand",
+  "Happy",
+  "Kind",
+  "Light",
+  "Noble",
+  "Pure",
+  "Quick",
+  "Rapid",
+  "Safe",
+  "True",
+  "Vivid",
+  "Warm",
+  "Wild",
+  "Wise",
+  "Zesty",
+  "Agile",
+  "Alert",
+  "Alive",
+  "Ample",
+  "Apt",
+  "Aware",
+  "Bliss",
+  "Brief",
+  "Chief",
+  "Civic",
+  "Clean",
+  "Crisp",
+  "Dense",
+  "Early",
+  "Easy",
+  "Elite",
+  "Equal",
+  "Exact",
+  "Extra",
+  "Fancy",
+  "Fine",
+  "Firm",
+  "First",
+  "Fleet",
+  "Fresh",
+  "Full",
+  "Glad",
+  "Gold",
+  "Good",
+  "Great",
+  "Green",
+  "Handy",
+  "Hardy",
+  "Ideal",
+  "Inner",
+  "Joint",
+  "Jolly",
+  "Just",
+  "Large",
+  "Lean",
+  "Legal",
+  "Level",
+  "Live",
+  "Local",
+  "Long",
+  "Loyal",
+  "Lucky",
+  "Major",
+  "Merry",
+  "Mild",
+  "Minor",
+  "Moral",
+  "Naval",
+  "Neat",
+  "New",
+  "Next",
+  "Nice",
+  "Open",
+  "Other",
+  "Outer",
+  "Owed",
+  "Peak",
+  "Plain",
+  "Plus",
+  "Polar",
+  "Proud",
+  "Ready",
+  "Real",
+  "Rich",
+  "Rigid",
+  "Round",
+  "Royal",
+  "Sleek",
+  "Slim",
+  "Snug",
+  "Solid",
+  "Sonic",
+  "Sound",
+  "Spare",
+  "Stark",
+  "Steep",
+  "Still",
+  "Super",
+  "Tidy",
+  "Tiny",
+  "Top",
+  "Total",
+  "Ultra",
+  "Upper",
+  "Urban",
+  "Valid",
+];
+
+const nouns = [
+  "Arc",
+  "Ash",
+  "Atom",
+  "Aura",
+  "Axis",
+  "Base",
+  "Beam",
+  "Bell",
+  "Bird",
+  "Bolt",
+  "Bond",
+  "Byte",
+  "Cape",
+  "Cast",
+  "Cave",
+  "Clay",
+  "Clip",
+  "Code",
+  "Coil",
+  "Coin",
+  "Core",
+  "Cove",
+  "Cube",
+  "Dale",
+  "Dart",
+  "Dawn",
+  "Deck",
+  "Dew",
+  "Dock",
+  "Dove",
+  "Drop",
+  "Drum",
+  "Dune",
+  "Dust",
+  "Echo",
+  "Edge",
+  "Elm",
+  "Fawn",
+  "Fern",
+  "Finn",
+  "Fire",
+  "Fizz",
+  "Flux",
+  "Foam",
+  "Font",
+  "Fork",
+  "Fort",
+  "Fuse",
+  "Gate",
+  "Gaze",
+  "Gear",
+  "Glow",
+  "Glen",
+  "Grit",
+  "Gulf",
+  "Gust",
+  "Halo",
+  "Harp",
+  "Hawk",
+  "Haze",
+  "Helm",
+  "Hill",
+  "Hive",
+  "Hook",
+  "Horn",
+  "Iris",
+  "Isle",
+  "Jade",
+  "Jazz",
+  "Kite",
+  "Knot",
+  "Lake",
+  "Lane",
+  "Lark",
+  "Leaf",
+  "Lens",
+  "Lime",
+  "Link",
+  "Lock",
+  "Loft",
+  "Loop",
+  "Loom",
+  "Lynx",
+  "Malt",
+  "Mane",
+  "Maze",
+  "Mesa",
+  "Mint",
+  "Mist",
+  "Moon",
+  "Moss",
+  "Muse",
+  "Nest",
+  "Node",
+  "Nova",
+  "Oaks",
+  "Onyx",
+  "Opus",
+  "Orb",
+  "Orca",
+  "Palm",
+  "Path",
+  "Peak",
+  "Pier",
+  "Pine",
+  "Plum",
+  "Pond",
+  "Port",
+  "Pulse",
+  "Quay",
+  "Raft",
+  "Rail",
+  "Rain",
+  "Reef",
+  "Rift",
+  "Ring",
+  "Rise",
+  "Rock",
+  "Root",
+  "Rose",
+  "Rune",
+  "Rush",
+  "Sage",
+  "Salt",
+  "Sand",
+  "Seal",
+  "Seed",
+  "Silk",
+  "Silo",
+  "Slate",
+  "Snow",
+  "Soar",
+  "Sole",
+  "Span",
+  "Spar",
+  "Star",
+  "Stem",
+  "Surf",
+  "Tarn",
+  "Tide",
+  "Tor",
+  "Vale",
+  "Vane",
+  "Vast",
+  "Vine",
+  "Volt",
+  "Wave",
+  "Webb",
+  "Well",
+  "Whim",
+  "Wick",
+  "Wind",
+  "Wing",
+  "Wire",
+  "Wolf",
+  "Wood",
+  "Wren",
+  "Yarn",
+  "Yew",
+  "Zinc",
+];
+
+export function generateName(): string {
+  const adj = adjectives[Math.floor(Math.random() * adjectives.length)];
+  const noun = nouns[Math.floor(Math.random() * nouns.length)];
+  return `${adj} ${noun}`;
+}
+
+export function toSlug(name: string): string {
+  return name
+    .toLowerCase()
+    .replace(/[^a-z0-9]+/g, "-")
+    .replace(/^-|-$/g, "");
+}
